Day 14 challenge in-progress
Part a works fine, but to finish part b the computation of the new polymer
string takes much too long based on size explosion.  Optimization of the
generator seems a dead end: a better approach is likely to interleave the
molecule counting and the generation; or, using a bottom-up technique with
precomputed molecule counting.