Day 15 challenge abandoned
Part a works fine.  Part b passed the test case, but fails on live data.
The dynamic programming cost matrix was coded to assume that moves only
happen to the right or down.  I assume for the live data the least risky
path includes a jog to the left or upward.  I'm not sure how to include
moving up or left given the starting position at the upper left.