Day 20 challenge abandoned.
Apparently, my idea of pre-padding the working image borders to simulate an
infinite board is broken because the value needed to reasonably fill the
unspecified cells depends on the first byte of the image enhancement
algorithm?

Even using that byte as a fill character, I get incorrect answers.

Moving on to a more fun Sunday evening activity.