Day 21 challenge abandoned.
Part a is complete.

Part b cannot be brute-forced, although that's what the code is doing now.
To solve this, I need to memoize the games, although I'm not sure if that
should be done using the start positions, the die history, the scores, or
some combination.  As usual, I can't be bothered :(