Day 22 challenge abandoned.
Part a is complete.

Part b cannot be brute-forced, although that's what the code is doing now.
To solve this, I think I need to use a BTree or something like it to keep
track of regions of reactor cube space, and the number of blocks within
that region.  Having one massive 3d array is not feasible.